rename validate path in structural tests
